refactor: 폴더 이름 변경, 테스트 패키지 이동, address 추가
