+++ backend/back/src/main/java/com/back/domain/welfare/center/center/properties/CenterApiProperties.java
@@ -1,4 +1,4 @@
-package com.back.domain.welfare.center.center.config;
+package com.back.domain.welfare.center.center.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
