+++ backend/back/src/test/java/com/back/domain/member/geo/service/GeoServiceTest.java
@@ -1,4 +1,4 @@
-package com.back.global.geo.service;
+package com.back.domain.member.geo.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -15,10 +15,8 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.back.domain.member.geo.dto.GeoApiResponseDto;
 import com.back.domain.member.geo.entity.AddressDto;
-import com.back.domain.member.geo.entity.GeoApiResponseDto;
-import com.back.domain.member.geo.service.GeoApiService;
-import com.back.domain.member.geo.service.GeoService;
 
 @ActiveProfiles("test")
 @SpringBootTest
